backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.1
openai==1.15.0
python-multipart==0.0.9